Removed Flutter notification stuff and cleaned up test notification files
